重构：STC32 任务框架模块化 + OpenART 持续发送
- STC32 拆分 control.c 为 line_ctrl(差速巡线)/ball_ctrl(舵机球稳)/task_sched(任务2/3/5/6框架)
- 新增 imu_ctrl：IMU660RA 硬件 SPI3，X 轴 yaw + 开机静止零点标定 + kd_yaw 阻尼
- main.c 改三态循环：菜单 ⇄ 任务 ⇄ 结果，任务结束 Key4 回菜单
- config 持久化 16→32 槽：新增 ball PID + 舵机标定 + kd_yaw，TAG/校验位后移
- OpenART 移除 START/STOP/CAL 命令门控，上电即持续发送 B/N 帧
- protocol proto_ball_valid 改 int16 匹配菜单 BOOL 项
- uvproj 注册新模块，.gitignore 排除构建日志/烧录工具

+++ OurProject/project/WcLibraries/protocol.c
@@ -22,7 +22,7 @@ static uint8 rx_idx;
 
 /* ── 全局状态 ── */
 int16 proto_ball_x      = 0;
-uint8 proto_ball_valid  = 0;
+int16 proto_ball_valid  = 0;
 
 /* ── 内部：解析一行 ── */
 static void parse_line(const char *line)
